libs/pyproject.toml
[tool.poetry]
name = "libs"
version = "1.0"
description = ""
authors = ["Tavin S Alderson <neot.oncloud@pm.me>"]

[tool.poetry.dependencies]
python = ">= 3.9"
confluent-kafka = {extras = ["avro", "schema-registry"], version = "^2.3.0"}
munch = "^4.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.ops.dependencies]
podman-compose = "^1.0.6"

[tool.pylint.format]
max-line-length = 100

[tool.pylint.'MESSAGES CONTROL']
disable = "C0413,C0114"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
